[expo-notifications] Scoped channels
